% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linalg.R
\name{matrix_sqrt}
\alias{matrix_sqrt}
\title{Matrix Square Root}
\usage{
matrix_sqrt(X, symmetric = FALSE)
}
\arguments{
\item{X}{A matrix}

\item{symmetric}{If \code{TRUE}, the matrix is assumed to be symmetric.
Passed to \code{eigen}.}
}
\value{
A matrix of the same shape as \code{X}.
}
\description{
Calculates the square root of a matrix by using its spectral decomposition.
}
