% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighbors.R
\name{mydist}
\alias{mydist}
\title{Euclidean Distance}
\usage{
mydist(x1, x2)
}
\arguments{
\item{x1, x2}{numeric vectors.}
}
\value{
the euclidean distance between the inputs.
}
\description{
Calculates the euclidean distance between two vectors
\code{x1} and \code{x2}.
}
