% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighbors.R
\name{nearest_neighbors}
\alias{nearest_neighbors}
\title{Nearest Neighbors}
\usage{
nearest_neighbors(X, x_query, k, dist_func = mydist, ...)
}
\arguments{
\item{X}{Feature matrix of shape [n_samples, n_features].}

\item{x_query}{A single vector of shape [n_features].}

\item{k}{The number of nearest neighbors to compute.}

\item{dist_func}{A function that defines a valid distance.
The default is euclidean distance.}

\item{...}{These parameters are passed into \code{dist_func}.}
}
\value{
The indices of the k-nearest neighbors.
}
\description{
Find the k nearest neighbors from a point \code{x_query}
and all rows in a dataset \code{X}.
}
