% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighbors.R
\name{predict_knn}
\alias{predict_knn}
\title{K-Nearest Neighbor Predictions}
\usage{
predict_knn(X, Y, x_query, k, dist_func = mydist, ...)
}
\arguments{
\item{X}{dataset of shape [n_samples, n_features]}

\item{Y}{target vector of shape [n_samples]}

\item{x_query}{vector of shape [n_features] to make predictions on.}

\item{k}{number of nearest neighbors to use for predictions.}

\item{dist_func}{function used to calculate the distance between two points.}

\item{...}{These arguments are passed into \code{dist_func}.}
}
\value{
A single prediction for \code{x_query}.
}
\description{
Make predictions on a data point \code{x_query}
based on the k-nearest neighbors in the dataset
\code{X}. Predictions are made by averaging the
\code{Y} values of the k-neighest neighbors.
}
