% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linalg.R
\name{sample_covariance}
\alias{sample_covariance}
\title{Sample Variance-Covariance Matrix}
\usage{
sample_covariance(X)
}
\arguments{
\item{X}{A matrix of shape [n_samples, n_features].}
}
\value{
The covariance matrix of \code{X}. Has shape [n_features, n_features].
}
\description{
Calculates the sample variance-covariance matrix using the biased MLE estimate.
}
